requirements.txt
streamlit
opencv-python-headless
numpy
pathlib
os-sys
pandas
matplotlib
pillow
tensorflow